requirement.txt
easyocr==1.6.2
matplotlib==3.7.1
opencv-python-headless==4.5.4.60
numpy==1.24.2
streamlit
pillow==9.5.0
langchain_groq
langchain_core
pydantic